003-remove-stopwords/requirements.txt
nltk==3.8.1
indic-nlp-library==0.92
